Change in function signature in cuda 8 vs cuda 9.1

+++ host-side/qemu-3.1.0/cudaforward.cu
@@ -43,7 +43,7 @@ extern "C" {
             char *deviceAddress,
             const char *deviceName,
             int ext,
-            size_t size,
+            int size,
             int constant,
             int global
             );
